imp: Optimized recipes rendering speed
